Improve error management in bookmark feature

+++ src/FBN/GuideBundle/Bookmark/BookmarkManager.php
@@ -25,36 +25,42 @@ class BookmarkManager
     /**
      * @var Session
      */
-    private $session;    
+    private $session;
 
     /**
      * @var PropertyAccessor
      */
-    protected $accessor;    
+    protected $accessor;
 
-	public function __construct(TokenStorageInterface $tokenStorage, EntityManager $em, Session $session)
-	{
+    public function __construct(TokenStorageInterface $tokenStorage, EntityManager $em, Session $session)
+    {
         $this->tokenStorage = $tokenStorage;
         $this->em = $em;
-        $this->session = $session;        
+        $this->session = $session;
         $this->accessor = PropertyAccess::createPropertyAccessor();
-	}
+    }
 
-    public function manage($bookmarkAction, $bookmarkId, $bookmarkEntite, $bookmarkEntiteId) {           
+    public function manage($bookmarkAction, $bookmarkId, $bookmarkEntite, $bookmarkEntiteId)
+    {
 
         // Only authorized actions and bookmarks
         if (!in_array($bookmarkAction, $this->session->get('bookmarkAction')) || !in_array($bookmarkId, $this->session->get('bookmarkId'))) {
-            return new JsonResponse('',403); 
-        }                        
+            return new JsonResponse('', 403);
+        }
 
-        if ($bookmarkAction == 'add') {                              
+        if ($bookmarkAction == 'add') {
 
             // Only authorized entities
             if (!in_array($bookmarkEntite, $this->session->get('bookmarkEntite')) || !in_array($bookmarkEntiteId, $this->session->get('bookmarkEntiteId'))) {
-                return new JsonResponse('',403); 
-            }                
-          
-            $favori = $this->add($bookmarkEntite, $bookmarkEntiteId);                
+                return new JsonResponse('', 403);
+            }
+
+            $favori = $this->add($bookmarkEntite, $bookmarkEntiteId);
+
+            if (null === $favori) {
+                return new JsonResponse('', 404);
+            }
+
             $bookmarkId = $favori->getId();
 
             $this->session->set('bookmarkAction', array('remove'));
@@ -63,9 +69,12 @@ class BookmarkManager
             return new JsonResponse(array('bookmarkAction' => 'remove',
                 'bookmarkId' => $bookmarkId,
             ));
-        }
-        elseif($bookmarkAction == 'remove') {
-            $this->remove($bookmarkId);
+        } elseif ($bookmarkAction == 'remove') {
+            $favori = $this->remove($bookmarkId);
+
+            if (null === $favori) {
+                return new JsonResponse('', 404);
+            }
 
             $bookmarkId = null;
 
@@ -74,49 +83,58 @@ class BookmarkManager
 
             return new JsonResponse(array('bookmarkAction' => 'add',
                 'bookmarkId' => $bookmarkId,
-            ));                    
-        }
-        elseif($bookmarkAction == 'remove_only') {
+            ));
+        } elseif ($bookmarkAction == 'remove_only') {
             $this->remove($bookmarkId);
 
             $bookmarkIds = $this->session->get('bookmarkId');
-            $bookmarkIdKey = array_search($bookmarkId, $bookmarkIds);                
+            $bookmarkIdKey = array_search($bookmarkId, $bookmarkIds);
             unset($bookmarkIds[$bookmarkIdKey]);
             $this->session->set('bookmarkId', $bookmarkIds);
 
-            return new JsonResponse(array('bookmarkId' => $bookmarkId));  
-        }        
+            return new JsonResponse(array('bookmarkId' => $bookmarkId));
+        }
     }
 
     private function add($entite, $entiteId)
-    {        
+    {
         // To access add function user connection is mandatory du to login entry point
         $user = $this->getUser();
 
         $entite = ucfirst($entite);
 
-        $repo = $this->em->getRepository('FBNGuideBundle:' . $entite);
+        $repo = $this->em->getRepository('FBNGuideBundle:'.$entite);
 
         $instance = $repo->findOneBy(array('id' => $entiteId));
 
+        if (null === $instance) {
+            return;
+        }
+
         $favori = new Favori();
         $favori->setUser($user);
-        $this->accessor->setValue($favori, $entite, $instance);        
+        $this->accessor->setValue($favori, $entite, $instance);
 
-        $this->em->persist($favori); 
+        $this->em->persist($favori);
         $this->em->flush();
 
-        return $favori;         
-    }  
+        return $favori;
+    }
 
     private function remove($bookmarkId)
     {
         $favori = $this->em->getRepository('FBNGuideBundle:Favori')
-            ->findOneBy(array('id' => $bookmarkId));            
+            ->findOneBy(array('id' => $bookmarkId));
+
+        if (null === $favori) {
+            return;
+        }
 
         $this->em->remove($favori);
         $this->em->flush();
-    }       
+
+        return $favori;
+    }
 
     public function checkStatus($entite, $entiteId)
     {
@@ -124,19 +142,18 @@ class BookmarkManager
         $bookmarkAction = 'add';
         $bookmarkId = null;
 
-        if(null !== $user = $this->getUser())
-        {
-	        if (is_object($user) || $user instanceof UserInterface) {            
-	            $userId = $user->getId();
+        if (null !== $user = $this->getUser()) {
+            if (is_object($user) || $user instanceof UserInterface) {
+                $userId = $user->getId();
 
-	            $favori = $this->em->getRepository('FBNGuideBundle:Favori')
-	                    ->getFavoriByEntiteId($userId, $entite, $entiteId);	                    
+                $favori = $this->em->getRepository('FBNGuideBundle:Favori')
+                        ->getFavoriByEntiteId($userId, $entite, $entiteId);
 
-	            if(null !== $favori) {
-	                $bookmarkAction = 'remove';
-	                $bookmarkId = $favori->getId();
-	            }
-	        }
+                if (null !== $favori) {
+                    $bookmarkAction = 'remove';
+                    $bookmarkId = $favori->getId();
+                }
+            }
         }
 
         $this->setSessionVariable(array($bookmarkAction), array($bookmarkId), array($entite), array($entiteId));
@@ -147,9 +164,9 @@ class BookmarkManager
     }
 
     public function setSessionVariable($bookmarkAction, $bookmarkId, $entite, $entiteId)
-    {        
+    {
         $this->session->set('bookmarkAction', $bookmarkAction);
-		$this->session->set('bookmarkId', $bookmarkId);        
+        $this->session->set('bookmarkId', $bookmarkId);
         $this->session->set('bookmarkEntite', $entite);
         $this->session->set('bookmarkEntiteId', $entiteId);
     }
@@ -158,11 +175,10 @@ class BookmarkManager
     {
         $securityToken = $this->tokenStorage->getToken();
 
-        if (null !== $securityToken)
-        {
+        if (null !== $securityToken) {
             return $securityToken->getUser();
         }
 
-        return null;
-    }    
+        return;
+    }
 }
